+++ gui/src/store.js
@@ -1,0 +1,10 @@
+// Zentraler Zustand der Anwendung
+import create from 'zustand';
+
+export const useStore = create((set) => ({
+  project: null,
+  images: [],
+  prefs: {},
+  setProject: (p) => set({ project: p }),
+  setImages: (imgs) => set({ images: imgs }),
+}));